src/requirements.txt
# python>=3.9
pandas
numpy
PyYAML
json
gitpython
flatdict